requirements.txt
openai-whisper
pydub
requests
iso8601
reportlab
torch
tqdm
flask
audioop-lts
